fixPrivateRoom
